+++ src/components/wordCounter.ts
@@ -33,7 +33,7 @@ export class WordCounter {
         this.setStatus()
     }
 
-    async counts(merge: boolean = true, file = this.extension.workshop.manager.rootFile()): Promise<TexCount> {
+    async counts(merge: boolean = true, file = this.extension.manager.rootFile): Promise<TexCount> {
         return new Promise((resolve, _reject) => {
             if (file === undefined) {
                 this.extension.logger.addLogMessage('A valid file was not give for TexCount')
@@ -47,7 +47,6 @@ export class WordCounter {
             args.push('-brief')
             let command = configuration.get('path') as string
             if (configuration.get('docker.enabled')) {
-                this.extension.workshop.manager.setEnvVar()
                 if (process.platform === 'win32') {
                     command = path.resolve(this.extension.extensionRoot, './scripts/countword-win.bat')
                 } else {
@@ -95,23 +94,6 @@ export class WordCounter {
                 }
             })
         })
-    }
-
-    async count(merge: boolean = true) {
-        const texCount = await this.counts(merge)
-        if (texCount.words.body) {
-            let floatMsg = ''
-            if (texCount.instances.floats > 0) {
-                floatMsg = `and ${texCount.instances.floats} float${
-                    texCount.instances.floats > 1 ? 's' : ''
-                } (tables, figures, etc.) `
-            }
-            vscode.window.showInformationMessage(
-                `There are ${texCount.words.body} words ${floatMsg}in the ${
-                    merge ? 'LaTeX project' : 'opened LaTeX file'
-                }.`
-            )
-        }
     }
 
     parseTexCount(text: string): TexCount {
@@ -182,12 +164,14 @@ export class WordCounter {
     async pickFormat() {
         const texCount = await this.counts()
 
-        const templates = ['${words} Words', '${headers} Headers', '${floats} Floats', '${math} Equations']
+        const templates = ['${words} Words', '${wordsBody} Words', '${headers} Headers', '${floats} Floats', '${math} Equations']
         const options: { [template: string]: string } = {}
         for (const template of templates) {
             options[template] = this.formatString(texCount, template)
+            if (template.startsWith('${wordsBody}')) {
+                options[template] += ' (body only)'
+            }
         }
-        options['custom'] = 'custom'
 
         const choice = await vscode.window.showQuickPick(Object.values(options), {
             placeHolder: 'Select format to use'